+++ packages/tables/src/lib/core/aggregateRenderers/Mean.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	/**
-	 * The `Mean` component renders the (arithemtic) mean of a set of values.
+	 * The `Mean` component renders the (arithmetic) mean of a set of values.
 	 * @component
 	 */
 
@@ -10,12 +10,17 @@
 	/**
 	 * Array of values to be displayed.
 	 */
-	export let values;
+	export let values: number[];
 
 	$: meanVal = mean(values);
 
 	export let formatString = '0.0f';
+
+	let f = format(formatString);
 	$: f = format(formatString);
+
+	// This suppresses warnings due to the RowRenderer providing props that aren't used.
+	$$restProps;
 </script>
 
-<span>{f(meanVal)}</span>
+<span>{meanVal === undefined ? 'undefined' : f(meanVal)}</span>