+++ packages/tables/src/lib/core/renderers/Dot.svelte
@@ -36,6 +36,9 @@
 	$: d = (value - extent[0]) / (extent[1] - extent[0]);
 
 	const fPercentage = format('0.0%');
+
+	// This suppresses warnings due to the RowRenderer providing props that aren't used.
+	$$restProps;
 </script>
 
 <div