+++ packages/tables/src/lib/table/Table.stories.svelte
@@ -8,6 +8,7 @@
 </script>
 
 <script lang="ts">
+	import type { ColSpec } from '$lib/core/lib/types';
 	import { Button, Input } from '@ldn-viz/ui';
 	import { Story, Template } from '@storybook/addon-svelte-csf';
 
@@ -59,7 +60,35 @@
 		]
 	};
 
-	let wideTableSpec = { columns: [] };
+	const tableSpecPartiallySortable = {
+		showColSummaries: false,
+		columns: [
+			{
+				short_label: 'first_name',
+				label: 'First Name',
+
+				cell: {
+					renderer: 'TextCell'
+				}
+			},
+
+			{
+				short_label: 'last_name',
+				label: 'Last Name',
+				sortable: false,
+				cell: { renderer: 'TextCell' }
+			},
+
+			{
+				short_label: 'pet',
+				label: 'Pet',
+				sortable: false,
+				cell: { renderer: 'TextCell' }
+			}
+		]
+	};
+
+	let wideTableSpec: { columns: ColSpec[] } = { columns: [] };
 	for (let i = 0; i < 25; i++) {
 		wideTableSpec.columns.push({
 			short_label: `field_${i}`,
@@ -76,14 +105,14 @@
 
 	let wideTableData = [];
 	for (let j = 0; j < 100; j++) {
-		let row = {};
+		let row: Record<string, number> = {};
 		for (let i = 0; i < 25; i++) {
 			row[`field_${i}`] = Math.round(Math.random() * 100) / 100;
 		}
 		wideTableData.push(row);
 	}
 
-	let dataSubset = [];
+	let dataSubset: Record<string, string | number>[] = [];
 	const randomlySelectRows = () => {
 		const selectedEntries = [];
 		const arrayCopy = [...data]; // Create a shallow copy of the input array
@@ -108,11 +137,15 @@
 
 <Story name="Table updates when data changes" source>
 	<Button on:click={randomlySelectRows}>Update</Button>
-	<Table data={dataSubset} {tableSpec} allowSorting />
+	<Table data={dataSubset} {tableSpec} allowSorting filename="My Table" />
 </Story>
 
 <Story name="Sortable Rows" source>
 	<Table {data} {tableSpec} allowSorting />
+</Story>
+
+<Story name="Sortable - but only on some columns" source>
+	<Table {data} tableSpec={tableSpecPartiallySortable} allowSorting />
 </Story>
 
 <Story name="Title" source>
@@ -129,14 +162,15 @@
 </Story>
 
 <Story name="Export buttons" source>
-	<Table {data} {tableSpec} exportBtns />
+	<Table {data} {tableSpec} dataDownloadButton imageDownloadButton />
 </Story>
 
 <Story name="Export buttons - relabel columns" source>
 	<Table
 		{data}
 		{tableSpec}
-		exportBtns
+		dataDownloadButton
+		imageDownloadButton
 		columnMapping={{ first_name: 'First Name', last_name: 'Last Name', pet: 'Pet' }}
 	/>
 </Story>