+++ packages/tables/src/lib/table/rows/groupRows/GroupHeadingRow.svelte
@@ -1,20 +1,21 @@
-<script>
+<script lang="ts">
 	import { ChevronDown, ChevronRight } from '@steeze-ui/heroicons';
 	import { Icon } from '@steeze-ui/svelte-icon';
 
+	import type { Group } from '../../../core/lib/types';
 	import GroupSizeBar from './GroupSizeBar.svelte';
 
 	export let group;
 	export let table;
 
-	const constructLabel = (group) => {
+	const constructLabel = (group: Group) => {
 		if (typeof group.maxRows !== 'undefined' && group.order.length > group.maxRows) {
 			return `${group.name} (${group.order.length} rows, ${group.order.length - group.maxRows} hidden)`;
 		}
 		return `${group.name} `;
 		//  return `${group.name} (${group.order.length} rows)`
 	};
-	const getGroupLevel = (name) => (name.match(new RegExp(' ∩ ', 'g')) || []).length;
+	const getGroupLevel = (name: string) => (name.match(new RegExp(' ∩ ', 'g')) || []).length;
 </script>
 
 <div class="flex was-tr">
@@ -42,7 +43,7 @@
 		{constructLabel(group)}
 	</div>
 
-	<!-- padding of equivalent siex to chevrons on other rows -->
+	<!-- padding of equivalent size to chevrons on other rows -->
 	{#each new Array(table.groupingFields.length - getGroupLevel(group.name)) as _i}
 		<!-- {@const g  = getGroup(group, i)} -->